requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
numpy==1.26.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
scikit_learn==1.3.2
torch==2.1.2
whisper==1.1.10
